Althea-main/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
google-genai>=1.0
lightgbm>=4.0
PyYAML>=6.0